[spreader][teams] Refactoring. Moved Sweeper class.
