+++ CodingDojo/server/src/main/java/com/codenjoy/dojo/services/multiplayer/Spreader.java
@@ -22,13 +22,11 @@ package com.codenjoy.dojo.services.multiplayer;
  * #L%
  */
 
-import com.codenjoy.dojo.services.Sweeper;
 import com.google.common.collect.LinkedHashMultimap;
 import com.google.common.collect.Multimap;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.List;
-import java.util.function.Predicate;
 import java.util.function.Supplier;
 
 import static java.util.stream.Collectors.toList;
